backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.100.2
python-dotenv==1.1.1
requests==2.32.4